ml_model/requirements.txt
openai-whisper==20231117
torch>=2.0.0
numpy>=1.24.0
soundfile>=0.12.1
ffmpeg-python>=0.2.0


